requirements.txt
django
djangorestframework
psycopg2
pandas
openpyxl
SQLAlchemy
django-cors-headers
